Ind. Backend V3
Created new Directories which have the ML model functions and implementations. More specifically, created a new .py file which creates an instance of the model used and loads the trained state dict of it. Then created a new function which transforms the image provided by the user to specific dimensions and passes it through the model in order to make a prediction. Model_dict dictionary, has the trained state dict of the model and ml_model.py has the functions used to perform a prediction on a give PIL type Image.