Individual frontend V8
Fixed mostly small UI features, such as NotebookCell component sizes on smaller screens, some background colors, and the links in both contents. The major difference is the implementation of an animation for the NavBar toggle button, which was made using CSS only. In addition changed the stle for the GitHub button when the navbar is collapsed.