Individual frontend V13
Fixed bug where if a user uploads an image, then click on the cancel button and then try to upload the SAME image again, nothing would show up in the UI, because React does not re-render the component since it knows the same image is uploaded. Fixed this bug by clearing up (emptying the input value) right after a user-uploaded image is provided as input.